+++ src/client/systems/player/System+TriggerPlayerShoot.cpp
@@ -19,7 +19,7 @@ namespace ECS {
         auto &keyboardEvent = eventManager->getEventsByType<Event::KeyboardEvent>();
         const auto size = aPos.size();
 
-        for (size_t i = 0; i < size; i++) {
+        for (unsigned short i = 0; i < size; i++) {
             if (!aType[i].has_value() || !aType[i].value().isPlayer || !aIsAlive[i].value().isAlive) {
                 continue;
             }
