+++ src/server/systems/enemy/System+EnemyShoot.cpp
@@ -19,7 +19,7 @@ namespace ECS {
         auto &world = Core::World::getInstance();
         const auto size = aPos.size();
 
-        for (size_t idx = 0; idx < size; idx++) {
+        for (unsigned short idx = 0; idx < size; idx++) {
             if (!aPos[idx].has_value() || !aType[idx].has_value() || !aIsAlive[idx].has_value()) {
                 continue;
             }
