+++ src/server/systems/network/System+ReceiveAknowledgment.cpp
@@ -18,11 +18,11 @@ namespace ECS {
         const auto size = events.size();
         std::vector<size_t> toRemove;
 
-        for (size_t i = 0; i < size; i++) {
+        for (unsigned short i = 0; i < size; i++) {
             auto &gameEvent = events[i];
 
             if (gameEvent.getType() == RType::ServerEventType::AKNOWLEDGMENT) {
-                size_t playerId = gameEvent.getEntityId();
+                unsigned short playerId = gameEvent.getEntityId();
 
                 if (aConnection[playerId].has_value()) {
                     auto &connection = aConnection[playerId].value();
